requirements.txt
pysolr
flask==0.11.1
cassandra-driver